consolidate jest files to simplify imports
