+++ src/slider/ColorInput.test.tsx
@@ -1,15 +1,9 @@
 import {act, cleanup, render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import {UserEvent} from '@testing-library/user-event/setup/setup';
-import React, {PropsWithChildren} from 'react';
-import {Provider} from 'react-redux';
-import {testForAccessibility} from '../jest/testForAccessibility';
-import store from '../redux/store';
+import React from 'react';
+import {ReduxContainer, testForAccessibility} from '../jest';
 import ColorInput from './ColorInput';
-
-const ReduxContainer = ({children}: PropsWithChildren) => {
-	return <Provider store={store}>{children}</Provider>;
-};
 
 describe('ColorInput.tsx', () => {
 	let user: UserEvent | undefined;
